requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
billiard==4.2.1
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dj-database-url==2.3.0
Django==5.0.7
django-celery-results==2.5.1
django-channels==0.7.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fcm-django==2.2.1
firebase-admin==6.5.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pip-check-reqs==2.5.3
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
razorpay==1.4.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
setuptools==73.0.1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.5.1
swapper==1.4.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.0.3
